move the vat files from under demo/ to under test/swingsetTests/
