+++ test/swingsetTests/contractHost/vat-fred.js
