+++ test/swingsetTests/gallery/vat-mint.js
