+++ test/swingsetTests/pixel-demo/bootstrap.js
