+++ test/swingsetTests/pixel-demo/vat-alice.js
