+++ test/swingsetTests/pixel-demo/vat-host.js
