+++ test/swingsetTests/pixel-demo/vat-mint.js
